refactor: logger
