+++ docs/cmd/hrp.md
@@ -22,4 +22,4 @@ Copyright 2021 debugtalk
 * [hrp har2case](hrp_har2case.md)	 - Convert HAR to json/yaml testcase files
 * [hrp run](hrp_run.md)	 - run API test
 
-###### Auto generated by spf13/cobra on 21-Dec-2021
+###### Auto generated by spf13/cobra on 22-Dec-2021
