+++ docs/cmd/hrp_boom.md
@@ -30,7 +30,6 @@ hrp boom [flags]
       --mem-profile-duration duration   Memory profile duration. (default 30s)
       --prometheus-gateway string       Prometheus Pushgateway url.
       --request-increase-rate string    Request increase rate, disabled by default. (default "-1")
-      --run-tasks string                Run tasks without connecting to the master, multiply tasks is separated by comma. Usually, it's for debug purpose.
       --spawn-count int                 The number of users to spawn for load testing (default 1)
       --spawn-rate float                The rate for spawning users (default 1)
 ```
@@ -39,4 +38,4 @@ hrp boom [flags]
 
 * [hrp](hrp.md)	 - One-stop solution for HTTP(S) testing.
 
-###### Auto generated by spf13/cobra on 21-Dec-2021
+###### Auto generated by spf13/cobra on 22-Dec-2021
