+++ docs/cmd/hrp_har2case.md
@@ -23,4 +23,4 @@ hrp har2case harPath... [flags]
 
 * [hrp](hrp.md)	 - One-stop solution for HTTP(S) testing.
 
-###### Auto generated by spf13/cobra on 21-Dec-2021
+###### Auto generated by spf13/cobra on 22-Dec-2021
