+++ docs/cmd/hrp_run.md
@@ -31,4 +31,4 @@ hrp run path... [flags]
 
 * [hrp](hrp.md)	 - One-stop solution for HTTP(S) testing.
 
-###### Auto generated by spf13/cobra on 21-Dec-2021
+###### Auto generated by spf13/cobra on 22-Dec-2021
